components/dispatch/requirements.txt
google-cloud-error-reporting
google-cloud-pubsub
google-cloud-storage
marshmallow
